pyproject.toml
[project]
name = "dynamic-extraction-engine"
version = "0.1.0"
description = "Config-driven multi-step LLM extraction engine with guardrails, cross-document validation, and credit decisioning"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "httpx>=0.27.0",
    "pydantic>=2.9.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
    "google-genai>=1.0.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.backends._legacy:_Backend"
